// rubiks-cube-fe/rubiks-cube-fe/src/utilities/enums.ts
export enum SideLabel {
  Front = "Front",
  Back = "Back",
  Left = "Left",
  Right = "Right",
  Top = "Top",
  Bottom = "Bottom",
}

export enum TileLabel {
  TL = "TopLeft",
  TM = "TopMiddle",
  TR = "TopRight",
  ML = "MiddleLeft",
  MM = "Middle",
  MR = "MiddleRight",
  BL = "BottomLeft",
  BM = "BottomMiddle",
  BR = "BottomRight",
}

export enum Colour {
  White = "White",
  Yellow = "Yellow",
  Blue = "Blue",
  Green = "Green",
  Red = "Red",
  Orange = "Orange",
}